Легкие правки 2

+++ 28.01.2023_C#1/21 дз/Program.cs
@@ -5,17 +5,17 @@
 
 Console.Clear();
 Console.Write("Введите X точки A: ");
-double X1 = double.Parse(Console.ReadLine());
+int X1 = int.Parse(Console.ReadLine());
 Console.Write("Введите Y точки A: ");
-double Y1 = double.Parse(Console.ReadLine());
+int Y1 = int.Parse(Console.ReadLine());
 Console.Write("Введите Z точки A: ");
-double Z1 = double.Parse(Console.ReadLine());
+int Z1 = int.Parse(Console.ReadLine());
 Console.Write("Введите X точки B: ");
-double X2 = double.Parse(Console.ReadLine());
+int X2 = int.Parse(Console.ReadLine());
 Console.Write("Введите Y точки B: ");
-double Y2 = double.Parse(Console.ReadLine());
+int Y2 = int.Parse(Console.ReadLine());
 Console.Write("Введите Z точки B: ");
-double Z2 = double.Parse(Console.ReadLine());
+int Z2 = int.Parse(Console.ReadLine());
 
 double d = Math.Sqrt(Math.Pow(X1 - X2, 2) + Math.Pow(Y1 - Y2, 2) +  Math.Pow(Z1 - Z2, 2));
 Console.WriteLine($"Длина отрезка = {d:f2}");
